BoriDjango-Server/requirements.txt
absl-py==1.3.0
altgraph==0.17.3
asgiref==3.6.0
astunparse==1.6.3
autopep8==2.0.1
bleach==6.0.0
build==0.10.0
cachetools==5.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
contourpy==1.0.6
cryptography==39.0.0
cycler==0.11.0
Django==4.1.5
djangorestframework==3.14.0
docutils==0.19
et-xmlfile==1.1.0
filelock==3.9.0
flatbuffers==22.12.6
fonttools==4.38.0
gast==0.4.0
google-auth==2.15.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.1
h5py==3.7.0
huggingface-hub==0.11.1
idna==3.4
importlib-metadata==6.0.0
jaraco.classes==3.2.3
jeepney==0.8.0
joblib==1.2.0
JPype1==1.4.1
keras==2.8.0
Keras-Preprocessing==1.1.2
keyring==23.13.1
kiwisolver==1.4.4
konlpy==0.6.0
libclang==14.0.6
lxml==4.9.2
Markdown==3.4.1
markdown-it-py==2.1.0
MarkupSafe==2.1.1
matplotlib==3.6.2
mdurl==0.1.2
more-itertools==9.0.0
mysqlclient==2.1.1
nltk==3.8.1
numpy==1.24.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
openpyxl==3.0.10
opt-einsum==3.3.0
packaging==22.0
pandas==1.5.2
Pillow==9.4.0
pkginfo==1.9.6
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
Pygments==2.14.0
pyinstaller==5.7.0
pyinstaller-hooks-contrib==2022.15
pyparsing==3.0.9
pyproject_hooks==1.0.0
python-dateutil==2.8.2
pytz==2022.7
PyYAML==6.0
readme-renderer==37.3
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rfc3986==2.0.0
rich==13.3.1
rsa==4.9
scikit-learn==1.2.0
scipy==1.10.0
seaborn==0.12.2
SecretStorage==3.3.3
sentence-transformers==2.2.2
sentencepiece==0.1.97
six==1.16.0
sqlparse==0.4.3
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-cpu==2.8.0
tensorflow-estimator==2.11.0
tensorflow-intel==0.0.1
tensorflow-io-gcs-filesystem==0.29.0
termcolor==2.1.1
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.1.0
tokenizers==0.13.2
tomli==2.0.1
torch==1.13.1
torchvision==0.14.1
tqdm==4.64.1
transformers==4.25.1
twine==4.0.2
typing_extensions==4.4.0
tzdata==2022.7
urllib3==1.26.13
webencodings==0.5.1
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.12.1
